Revert separate experiment_config package
