+++ model/standalone_driver/pyproject.toml
@@ -30,7 +30,6 @@ dependencies = [
   "icon4py-atmosphere-dycore~=0.2.0",
   "icon4py-atmosphere-diffusion~=0.2.0",
   "icon4py-common[io]~=0.2.0",
-  "icon4py-testing~=0.2.0",  # TODO(): remove this dependency when driver is fully standalone
   # external dependencies
   "typer>=0.20.0",
   "devtools>=0.12",
