+++ pyproject.toml
@@ -97,7 +97,6 @@ dependencies = [
   "icon4py-atmosphere-muphys~=0.2.0",
   "icon4py-common~=0.2.0",
   "icon4py-driver~=0.2.0",
-  "icon4py-experiment_config~=0.2.0",
   "icon4py-standalone_driver~=0.2.0"
 ]
 description = 'ICON model in Python.'
@@ -181,7 +180,6 @@ files = [
   # '$MYPY_CONFIG_FILE_DIR/model/driver/tests/',
   '$MYPY_CONFIG_FILE_DIR/model/testing/src/',
   # '$MYPY_CONFIG_FILE_DIR/model/testing/tests/',
-  '$MYPY_CONFIG_FILE_DIR/model/experiment_config/src/',
   '$MYPY_CONFIG_FILE_DIR/tools/src/',
   '$MYPY_CONFIG_FILE_DIR/bindings/src/',
   '$MYPY_CONFIG_FILE_DIR/model/standalone_driver/src',
@@ -197,7 +195,6 @@ $MYPY_CONFIG_FILE_DIR/model/atmosphere/subgrid_scale_physics/microphysics/src:
 $MYPY_CONFIG_FILE_DIR/model/atmosphere/subgrid_scale_physics/muphys/src:
 $MYPY_CONFIG_FILE_DIR/model/common/src:
 $MYPY_CONFIG_FILE_DIR/model/driver/src:
-$MYPY_CONFIG_FILE_DIR/model/experiment_config/src:
 $MYPY_CONFIG_FILE_DIR/model/standalone_driver/src:
 $MYPY_CONFIG_FILE_DIR/model/standalone_driver/tests:
 $MYPY_CONFIG_FILE_DIR/model/testing/src:
@@ -276,7 +273,6 @@ pythonpath = [
   "model/atmosphere/subgrid_scale_physics/muphys",
   "model/common",
   "model/driver",
-  "model/experiment_config",
   "model/standalone_driver",
   "model/testing",
   "tools"
@@ -345,11 +341,11 @@ typing-modules = ['gt4py.eve.extended_typing']
 unfixable = []
 
 [tool.ruff.lint.flake8-bugbear]
-# DeltaT / NumTimeSteps are aliases of immutable types (timedelta / int), so calling them
+# RelativeTime / NumTimeSteps are aliases of immutable types (timedelta / int), so calling them
 # in argument or dataclass defaults is safe (B008, RUF009).
 extend-immutable-calls = [
-  "icon4py.model.common.io.io.DeltaT",
-  "icon4py.model.common.io.io.NumTimeSteps"
+  "icon4py.model.common.time.NumTimeSteps",
+  "icon4py.model.common.time.RelativeTime"
 ]
 
 [tool.ruff.lint.isort]
@@ -428,7 +424,6 @@ icon4py-atmosphere-muphys = {workspace = true}
 icon4py-bindings = {workspace = true}
 icon4py-common = {workspace = true}
 icon4py-driver = {workspace = true}
-icon4py-experiment_config = {workspace = true}
 icon4py-standalone_driver = {workspace = true}
 icon4py-testing = {workspace = true}
 icon4py-tools = {workspace = true}
@@ -443,7 +438,6 @@ members = [
   "model/atmosphere/subgrid_scale_physics/muphys",
   "model/common",
   "model/driver",
-  "model/experiment_config",
   "model/standalone_driver",
   "model/testing",
   "tools"
